% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pend.R
\name{get}
\alias{get}
\title{Sets up pending Paxos-get operations.}
\usage{
get(...)
}
\arguments{
\item{...}{Keys to get.}
}
\value{
Accumulated vector of keys for pending Paxos-get operations.
}
\description{
Sets up pending Paxos-get operations.
}
