% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pend.R
\name{pend.set}
\alias{pend.set}
\title{Sets up pending Paxos-set operations.}
\usage{
pend.set(...)
}
\arguments{
\item{...}{Pairs of names and values to set.}
}
\value{
Accumulated list of names and values for pending Paxos-set
  operations.
}
\description{
Sets up pending Paxos-set operations.
}
\seealso{
[new.env()]
}
